kps toimg aftr usroupd8s usrnm ts p

+++ src/user/UserService.js
@@ -94,10 +94,14 @@ const FileService = require('../file/FileService');
         
         const user = await User.findOne({ where: { id: id } });
         user.username = updatedBody.username;
-        if(user.image){
-            await FileService.deleteProfileImage(user.image);
+        if(updatedBody.image){
+
+            if(user.image){
+                await FileService.deleteProfileImage(user.image);
+            }
+            user.image = await FileService.saveProfileImage(updatedBody.image);
+
         }
-        user.image = await FileService.saveProfileImage(updatedBody.image);
         await user.save();
         return{
             id: id,
